GEODE-3616: Clean up ClientHealthMonitorJUnitTest - renamed ClientHealthMonitorJUnitTest to ClientHealthMonitorIntegrationTest - added new ClientHealthMonitorJUnitTest
