+++ geode-core/src/test/java/org/apache/geode/internal/cache/tier/sockets/ClientHealthMonitorSelectorIntegrationTest.java
@@ -14,11 +14,9 @@
  */
 package org.apache.geode.internal.cache.tier.sockets;
 
-import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
 import org.apache.geode.test.junit.categories.ClientServerTest;
-import org.apache.geode.test.junit.categories.FlakyTest;
 import org.apache.geode.test.junit.categories.IntegrationTest;
 
 /**
@@ -27,7 +25,7 @@ import org.apache.geode.test.junit.categories.IntegrationTest;
  *
  */
 @Category({IntegrationTest.class, ClientServerTest.class})
-public class ClientHealthMonitorSelectorJUnitTest extends ClientHealthMonitorJUnitTest {
+public class ClientHealthMonitorSelectorIntegrationTest extends ClientHealthMonitorIntegrationTest {
   protected int getMaxThreads() {
     return 2;
   }
